Add class-wise thresholds

+++ perception/autoware_lidar_centerpoint/include/autoware/lidar_centerpoint/centerpoint_config.hpp
@@ -27,7 +27,7 @@ public:
     const std::size_t class_size, const float point_feature_size, const std::size_t cloud_capacity,
     const std::size_t max_voxel_size, const std::vector<double> & point_cloud_range,
     const std::vector<double> & voxel_size, const std::size_t downsample_factor,
-    const std::size_t encoder_in_feature_size, const std::vector<float> &score_thresholds,
+    const std::size_t encoder_in_feature_size, const std::vector<double> &score_thresholds,
     const float circle_nms_dist_threshold, const std::vector<double> yaw_norm_thresholds,
     const bool has_variance)
   {
